requirements.txt
pillow
rich
python-i18n[YAML]